pyproject.toml
[project]
name = "bot_yerdna"
version = "0.1.0"
description = "Bot Yerdna je AI agent"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "chainlit==2.0.dev0",
    "langchain-groq>=0.2.0",
    "langchain-openai>=0.2.2",
    "langchain>=0.3.3",
    "openai==1.51.2",
    "plotly>=5.24.1",
    "websockets==12.0",  # Pinned websockets version
    "together>=1.0",
    "tavily-python>=0.3",
]

[tool.setuptools]
packages = ["apka"]